+++ intel_extension_for_transformers/llm/runtime/graph/core/layers/layers.h
@@ -13,3 +13,5 @@
 //  limitations under the License.
 #include "ele_wise.h"
 #include "ele_reduce.h"
+
+#include "conv.h"
