+++ intel_extension_for_transformers/llm/runtime/graph/core/ne.h
@@ -55,6 +55,40 @@
 #define NE_ALIGNMENT 32
 #endif
 
+#define NE_ASSERT(x)                                                     \
+  do {                                                                   \
+    if (!(x)) {                                                          \
+      fprintf(stderr, "NE_ASSERT: %s:%d: %s\n", __FILE__, __LINE__, #x); \
+      abort();                                                           \
+    }                                                                    \
+  } while (0)
+
+//
+// logging
+//
+
+#define NE_DEBUG 0
+
+#if (NE_DEBUG >= 1)
+#define NE_PRINT_DEBUG(...) printf(__VA_ARGS__)
+#else
+#define NE_PRINT_DEBUG(...)
+#endif
+
+#if (NE_DEBUG >= 5)
+#define NE_PRINT_DEBUG_5(...) printf(__VA_ARGS__)
+#else
+#define NE_PRINT_DEBUG_5(...)
+#endif
+
+#if (NE_DEBUG >= 10)
+#define NE_PRINT_DEBUG_10(...) printf(__VA_ARGS__)
+#else
+#define NE_PRINT_DEBUG_10(...)
+#endif
+
+#define NE_PRINT(...) printf(__VA_ARGS__)
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -178,6 +212,26 @@ struct ne_init_params {
   size_t mem_size;   // bytes
   void* mem_buffer;  // if NULL, memory will be allocated internally
   bool no_alloc;     // don't allocate memory for the tensor data
+};
+
+//
+// compute types
+//
+
+enum ne_task_type {
+  NE_TASK_INIT = 0,
+  NE_TASK_COMPUTE,
+  NE_TASK_FINALIZE,
+};
+
+struct ne_compute_params {
+  enum ne_task_type type;
+
+  int ith, nth;
+
+  // work buffer for all threads
+  size_t wsize;
+  void* wdata;
 };
 
 #ifdef __cplusplus
